+++ app/Http/Controllers/MendudukijabatanController.php
@@ -156,8 +156,11 @@ class MendudukijabatanController extends Controller
     public function destroy(int $id)
     {
         $data = Mendudukijabatan::firstWhere('id', $id);
-        // Hapus Berkas Lama
-        unlink(public_path('file/mendudukijabatan/').$data->file);
+        // Cek apakah ada berkas?
+        if ($data->file !== null) {
+            // Hapus Berkas Lama
+            unlink(public_path('file/mendudukijabatan/').$data->file);
+        }
         $data->delete();
 
         return redirect()->route('mendudukijabatan.index')->with('message', 'Data Berhasil Dihapus!');
